+++ .gitignore
@@ -3,3 +3,4 @@
 .mypy_cache
 __pycache__
 .pytest_cache
+dist
