+++ pyproject.toml
@@ -15,7 +15,7 @@
 #
 
 [tool.poetry]
-name = "conf_parser"
+name = "addonfactory_splunk_conf_parser_lib"
 version = "0.3.1"
 description = "Splunk .conf files parser"
 authors = ["Splunk <addonfactory@splunk.com>"]
